backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
groq==0.15.0
PyPDF2==3.0.1
pytesseract==0.3.13
Pillow==11.1.0
python-docx==1.1.2
youtube-transcript-api==1.2.4
slowapi==0.1.9
python-dotenv==1.0.1
pydantic==2.10.5
pydantic-settings==2.7.1
tenacity==9.0.0
